requirements.txt
openai
pydantic>=2.0
streamlit
sqlmodel
python-dotenv
pytest
pytest-asyncio
